broke out globalstyles, did a bit of apollo boilerplate, getting ready to test that frontend and backend and speak to each other
