+++ frontend/package.json
@@ -71,7 +71,8 @@
     "env": {
       "development": {
         "presets": [
-          "next/babel"
+          "next/babel",
+          "@babel/preset-typescript"
         ],
         "plugins": [
           [
