+++ frontend/pages/_app.tsx
@@ -1,61 +1,49 @@
 import App, { Container } from 'next/app';
 import Page from '../components/Page';
-import { createGlobalStyle, ThemeProvider } from 'styled-components';
+import { GlobalStyle, theme } from '../components/styles/GlobalStyles';
+import { ThemeProvider } from 'styled-components';
+import { ApolloProvider } from 'react-apollo';
+import withData from '../lib/withData';
 
-const theme = {
-  red: '#FF0000',
-  black: '#393939',
-  grey: '#3A3A3A',
-  lightgrey: '#E1E1E1',
-  offwhite: '#EDEDED',
-  maxWidth: '1000px',
-  bs: '0 12px 24px 0 rgba(0, 0, 0, 0.09)'
+interface Props {
+  apollo: any
 };
 
-const GlobalStyle = createGlobalStyle`
-  @font-face {
-    font-family: 'radnika_next';
-    src: url('/static/radnikanext-medium-webfont.woff2')
-    format('woff2');
-    font-weight: normal;
-    font-style: normal;
+class MyApp extends App<Props> {
+static async getInitialProps({ Component, ctx}) {
+  let pageProps = {
+    query: null
+  };
+  if(Component.getInitialProps) {
+    pageProps = await Component.getInitialProps(ctx);
   }
-  html {
-    box-sizing: border-box;
-    font-size: 10px;
-  }
-  *, *:before, *:after {
-    box-sizing: inherit
-  }
-  body {
-    padding: 0;
-    margin: 0;
-    font-size: 1.5rem;
-    line-height: 2;
-    font-family: 'radnika_next';
-  }
-  a {
-    text-decoration: none;
-    color: ${theme.black};
-  }
-`;
+  pageProps.query = ctx.query;
+  return  { pageProps };
+}
 
-export default class MyApp extends App {
   render(){
-    const { Component } = this.props;
+    const { Component, apollo, pageProps } = this.props;
 
     return (
       <Container>
-        <GlobalStyle/>
-        <ThemeProvider
-          theme={theme}
+        <ApolloProvider
+          client={apollo}
         >
-          <Page>
-            <Component/>
-          </Page>
+          <GlobalStyle/>
+          <ThemeProvider
+            theme={theme}
+          >
+            <Page>
+              <Component
+                { ...pageProps }
+              />
+            </Page>
 
-        </ThemeProvider>
+          </ThemeProvider>
+        </ApolloProvider>
       </Container>
     )
   }
 }
+
+export default withData(MyApp);
